+++ config/application.rb
@@ -98,9 +98,6 @@ module RLetters
 
       # Disable request forgery protection in test environment
       config.action_controller.allow_forgery_protection = false
-
-      # Tell Action Mailer not to deliver emails to the real world
-      config.action_mailer.delivery_method = :test
     end
   end
 end